lustre: uapi: fix building fail against Power9 little endian
We use "%ll[dux]" for __u64 variable as an input/output modifier,
this may cause building error on some architectures which use "long"
for 64-bit types, for example, Power9 little endian.
Here add necessary typecasting (long long/unsigned long long) to
make the build correct.

WC-bug-id: https://jira.whamcloud.com/browse/LU-12705
Lustre-commit: 4eddf36ac360 ("LU-12705 build: fix building fail against Power9 little endian")
Reviewed-on: https://review.whamcloud.com/36007

+++ include/uapi/linux/lustre/lustre_user.h
@@ -521,7 +521,7 @@ struct lu_extent {
 } __packed;
 
 #define DEXT "[%#llx, %#llx)"
-#define PEXT(ext) (ext)->e_start, (ext)->e_end
+#define PEXT(ext) (unsigned long long)(ext)->e_start, (unsigned long long)(ext)->e_end
 
 static inline bool lu_extent_is_overlapped(struct lu_extent *e1,
 					    struct lu_extent *e2)
